refactor(mesaService): cria interface e classe de impl para o service da mesa
