+++ src/main/java/com/sameul/sistemarestauranteapi/mesa/service/MesaService.java
@@ -8,28 +8,8 @@ import com.sameul.sistemarestauranteapi.common.exceptions.ObjectNotFoundExceptio
 import com.sameul.sistemarestauranteapi.restaurante.enums.RestauranteErrorCode;
 import org.springframework.stereotype.Service;
 
-@Service
-public class MesaService {
 
-    private final MesaRepository repository;
-
-    public MesaService(MesaRepository repository){
-        this.repository = repository;
-    }
-
-    public void salvar(MesaRequest request){
-        try {
-            Mesa mesa = new Mesa(request.getStatus(), request.getRestaurante());
-            repository.save(mesa);
-        }catch (Exception exception){
-            throw new RuntimeException(exception.getMessage());
-        }
-    }
-    public void alterarStatusMesa(int mesaId, MesaStatus status){
-        Mesa mesa = repository.findById(mesaId)
-                .orElseThrow(() -> new ObjectNotFoundException(RestauranteErrorCode.RESTAURANTE_NAO_ENCONTRADO));
-
-        mesa.setStatus(status);
-        repository.save(mesa);
-    }
+public interface MesaService {
+    void salvar(MesaRequest request);
+    void alterarStatusMesa(int mesaId, MesaStatus status);
 }
